chore: added helper script
