--- scripts/open_permalinks.py
@@ -1,38 +1,0 @@
-"""Open permalinks generated by `hugo list all > my_file.csv`"""
-
-from tqdm import tqdm
-import csv
-from multiprocessing import Pool
-from pathlib import Path
-import webbrowser
-import fire
-import logging
-import urllib
-import urllib.request
-
-
-def main(path: Path, localhost: bool = False):
-
-    # load csv
-    with open(path, "r", encoding="utf-8") as f:
-        reader = csv.DictReader(f)
-        links = [row["permalink"] for row in reader]
-    logging.info(f"Found {len(links)} links")
-
-    logging.info("Checking links...")
-    with Pool() as pool:
-        pool.map(check_link, tqdm(links))
-
-
-
-def check_link(link: str):
-    status = urllib.request.urlopen(link).status
-
-    if status != 200:
-        logging.error(f"{link} | {status}")
-
-
-if __name__ == "__main__":
-    logging.basicConfig(level=logging.INFO)
-    # fire.Fire(main)
-    main("permalinks.csv")
